Extract shared caps-discovery helpers for Newznab/Torznab
buildCapsUrlCandidates, DEFAULT_GAME_CATEGORIES/_IDS, and
CAPS_DISCOVERY_TIMEOUT_MS were hand-duplicated between newznab.ts and
torznab.ts and had already drifted (10s vs 30s timeout for the
identical operation). Moved them into server/indexer-caps.ts and
reconciled the timeout to 15s. The getCategories retry loop itself
stays per-protocol -- it differs in real ways (SSRF vs enabled check,
headers, error-text extraction, logger) not worth forcing through one
shared path.
